- linux: fix build [TESTS] - build 6.2 branch

+++ backend/wbpublic/grtui/grtdb_object_filter.cpp
@@ -190,11 +190,11 @@ void DBObjectFilterFrame::refresh(int object_list_selection, int mask_list_selec
   _exclude_model->refresh();
   
   refill_list(_object_list, _model);
-  if (object_list_selection > -1 && object_list_selection < _model->count())
+  if (object_list_selection > -1 && (size_t)object_list_selection < _model->count())
     _object_list.set_selected(object_list_selection);
 
   refill_list(_mask_list, _exclude_model);
-  if (mask_list_selection > -1 && mask_list_selection < _exclude_model->count())
+  if (mask_list_selection > -1 && (size_t)mask_list_selection < _exclude_model->count())
     _mask_list.set_selected(mask_list_selection);
 
   std::stringstream out;
